refactor: aplica uso do modelo medgemma via ollama e altera versao do google adk para uso do adk ui
