+++ nutrixpert/__init__.py
@@ -1,0 +1,1 @@
+from .agent import agent
