+++ requirements.txt
@@ -22,6 +22,7 @@ colorama==0.4.6
 coloredlogs==15.0.1
 cryptography==45.0.7
 dataclasses-json==0.6.7
+Deprecated==1.2.18
 distro==1.9.0
 docstring_parser==0.17.0
 dotenv==0.9.9
@@ -33,7 +34,7 @@ filelock==3.19.1
 flatbuffers==25.2.10
 frozenlist==1.7.0
 fsspec==2025.9.0
-google-adk==1.14.0
+google-adk==0.3.0
 google-api-core==2.25.1
 google-api-python-client==2.181.0
 google-auth==2.40.3
@@ -115,6 +116,7 @@ opentelemetry-sdk==1.37.0
 opentelemetry-semantic-conventions==0.58b0
 orjson==3.11.3
 overrides==7.7.0
+packaging==24.2
 pandas==2.3.2
 pdfminer.six==20250506
 pdfplumber==0.11.7
@@ -190,6 +192,7 @@ watchdog==6.0.0
 watchfiles==1.1.0
 websocket-client==1.8.0
 websockets==15.0.1
+wrapt==1.17.3
 yarl==1.20.1
 zipp==3.23.0
 zstandard==0.25.0
